+++ scheduler/src/admin/jobs/hello.js
@@ -1,0 +1,5 @@
+import { appLogger } from '../../utils/index.js'
+
+export const hello = async job => {
+  appLogger.info({ description: 'Hello' })
+}
